Copy subject binaries into kernel using objcopy
Instead of linking the subject object files into the kernel image, use objcopy
to add the subject binaries as new sections.

With this change the subjects and the kernel are built and linked independently.
Thus the subject images are self-contained and do not use any kernel memory
apart from the stack.