Add initial τ0 subject
The τ0 subject is a privileged subject which works in cooperation with the
kernel. It communicates with the kernel via special interface pages, which are
visible to τ0 and the kernel. For now there is one page (τ0 -> kernel), which is
writable by τ0 and readable by the kernel.

Currently it instructs the kernel to switch between the two configured
scheduling plans.