Let config tool generate subject XML spec
Instead of writing an Ada spec file containing all subjects defined in
the policy, the skconfig tool now takes the binary to analyze from the
command line and writes a XML specification for it.

This avoids dependencies between the kernel, policy and config projects
and makes skconfig an independent tool.