Policy: Include generated initial subject state spec
To ease development the skconfig tool can be used to generate the
initial subject state XML specification from the subject (ELF) binary.

The generated specification can in turn be included in the final system
policy. This guarantees that initial subject state information is
automatically updated when the binary changes (i.e. stack address and
entry point).

The use of the skconfig tool is optional, initial subject state can also
be hard-coded in the policy (e.g. once the development has been
completed).