Trampoline: Jump to the AP entry point
The AP trampoline loads a temporary GDT with 32 bit segment descriptors
and switches to 32 bit protected mode. It then jumps to the AP entry
point defined in init.S.

This makes the final jump into the main kernel work. But since the APs have
no separate stack (yet), the result is not that pretty.