Add HPET capability register type
The type matches the specification of the general capability and id register as
defined by the IA-PC HPET specification Rev 1.0. This allows to constrain the
counter clock period value to the range of the COUNTER_CLK_PERIOD field, see
section 2.3.4 of the document.

Quote:

 "A value of 0 in this field is not permitted. The value in this field must be
  less than or equal to 05F5E100h (10^8 femptoseconds = 100 nanoseconds)."

This enables the simplifier to prove that there is no division by zero when
using the counter period to convert a duration to counter ticks (e.g. to pass to
the Set_Timer procedure).

Note: The quote is taken as is even though it should say "femtoseconds".