Set VM-exit 'ack interrupt on exit' control flag
With this flag set, the logical processor acknowledges the interrupt controller
and acquires the interrupt's vector number. This allows to to invoke the IRQ
handler directly instead of enabling interrupts and vectoring the event via the
IDT.

This setup effectively allows to handle external interrupts without extending
the IDT with ISR entries for all possible external interrupts.