Preserve pending events when setting subject state
Since the whole subject state is copied at the beginning of the VMX exit handler
and written back at the end, a pending event might get lost. On an exit due to
an external interrupt for the current subject, the IRQ handler would set the
pending event field correctly, but it would be overwritten shortly after.