Rework dumper
The dumper is now interrupt-driven. The injected interrupt identifies the
subject which is to be processed by the dumper. The mapping of interrupt vectors
to subject ids is as follows: Subject id = Vector - 32.

All the processing is done in the interrupt handler since a dump of a subject
state must be finished before the next dump can begin.