VT: Use framebuffer array instead of 'Address clause
The previous VGA console update process could be interrupted while
re-declaring the active framebuffer address. This led to artifacts on
the screen when typing on the keyboard.

Fix this by using an array of framebuffers. Furthermore, make the
Current_Slot variable atomic to guarantee that updating it is not
interrupted.