skpolicy: Add EPT_Pointer to generated subject spec
The new EPT_Pointer record field stores a subjects Extended-Page-Table
Pointer (EPTP), as specified by Intel SDM Vol. 3C, chapter 24.5.11.

If a subject's profile is native, the EPTP is set to 0 and the PML4
address is written as usual. For a subject with profile VM, the EPTP is
generated using the PML4 address and the EPTP flags. The subject PML4
address is set to 0.

The EPTP flags constant specifies, that the EPT paging-structure memory
type is WB, the EPT page-walk length is 4. The 'accessed/dirty' flags
for EPT are not enabled, since this feature is not supported by Ivy
Bridge processors.

Having both PML4 address and EPTP fields simplifies the kernel code,
since it can unconditionally write the given values to the corresponding
VMCS fields. The kernel does not need to determine if a given subject
uses EPT or not.