Set interrupt-window exiting on skipped event injection
Set the interrupt-window exiting VM processor control, if a pending
event exists but the injection must be postponed because the subject is
currently not ready to handle interrupts.

A VM exit will occur as soon as the subject is in an interruptible state
so the pending interrupt can be injected with as little delay as
possible.