Scheduler: Send IPI on inter-core signal
The Handle_Hypercall procedure sends an IPI to the CPU core of the
destination subject if the signal type is 'synchronous'.

The Handle_Irq handler of the target core ignores the special IPI IRQ
(254) and continues scheduling subjects. This mechanism is used to
immediately inject an interrupt into a running subject.