Kernel: Add support for multiple subject events
The SK.Subjects package now implements a per-subject event queue of
maximal 32 pending events. If a subject does not consume its events and
the queue runs full, a (debug) message is displayed and new events are
ignored.

Since multiple logical CPUs manipulate the event queues concurrently, it
is protected by a spinlock.