Rename 'Signals' to 'Events'
A subject can specify an optional event table which contains two kind of
events:

A handover event transfers execution to a destination subject, optionally
injecting an interrupt with a given vector number. The policy tool
enforces that the destination subject must run on the same logical CPU
as the triggering subject.

An interrupt event injects an interrupt with specified vector to a
destination subject. If the send_ipi attribute is set to true, an
inter-processor interrupt (IPI) is sent to the logical CPU executing the
destination subject. This mechanism leads to the immediate injection of
the specified vector into the destination subject.

The policy tool enforces that the destination subject for an interrupt
event with IPI must run on a different logical CPU than the triggering
subject.