Crypter: Implement sender and receiver packages
The crypter service is comprised of a request and a response page. The
request page is located at physical address 0x29d000 and the
response page at 0x29e000.

Processing is done as follows: the client fills his data into a message
and copies that message to the crypter request page. It then notifies
the crypter by sending an event.

When the crypter receives an event, it copies the request message,
processes it and writes the response to the crypter response page. The
crypter then signals completion by sending an event to the requesting
client.

Currently the crypter simply echos back the request as response.