Add initial thesis report document
The report is titled 'Muen - A high assurance separation kernel for
the Intel x86 architecture'.

We choose to name the kernel 'Muen' which is Japanese and translates to
'unrelated/without relation', an allegory of the separated subjects
running on the kernel.

The root of the word Muen is 'Mu' which denotes a negative: the
absence of something. It is a keyword in Chan and Zen Buddhism and also
part of the Jargon File [1].

[1] - http://www.catb.org/jargon/html/M/mu.html