Implement bitfield-based event processing
The new SK.Events package provides an interface to insert and consume
subject events.

Pending events are internally stored in bitfields which are updated
atomically. This design is much simpler than the event list and avoids
the usage of spinlocks.

The mechanism is modeled after the APICv posted-interrupt processing
described in Intel SDM, Vol. 3C, section 29.6.