Patch up bzImage 64bit entry point
The bzImage entry point code first tries to reload some
segment selectors. Unfortunately we never load a real table
into VMs, and building a new table type just for 5 instructions
seemed excessive. So pattern match and skip the code - it's unlikely
to change often and it's written in assembly (so pattern matching won't
break with new compiler versions).

The bootparams flag to change the behaviour also changes behavior in an
undesirable way elsewhere in the kernel.