Kernel: Assign continuous CPU IDs during init
Upon system init each CPU atomically increments a global CPU counter and
assigns itself a unique CPU ID. The ID is stored in the per-CPU storage
area where it is accessible from SPARK.

The value of the variable can only be written once paging is active. Thus the
CPU ID is kept in the (r/)esi register until the per-CPU storage area is
available.