Replace assembly by synthetic SPARK code for proof
SPARK 2014 does not know assembly. While assembly code called from SPARK can be
hidden by SPARK_Mode => Off, an assembly entry has to be "simulated" by a SPARK
main program which approximates the behaviour.

When a suitable model of the userspace is added to init.adb, global properties
of Muen can be shown.