Place VMXON, VMCS and kernel regions at fixed locations
VMXON and VMCS regions have to reside below 1MB, as they are initialized in
init.S before paging is enabled. Consequently, those regions don't have to be
mapped into the kernel and no complicated static tables need to be present at
initialization time. This is no severe limitation, as Tau0 manages VMCS
regions (and possibly VMXON regions for AP's) in the future - only the BSP's
VMXON region and Tau0's own VMCS will remain below 1MB then.

The kernel regions need to be placed and mapped as defined kernel/kernel.ld.

The RAM between 0 and 640kB is defined as non-allocatable now, i.e. mucfgalloc
disregards it and only fixed regions are placed there. The previous approach
that would use the @below attribute to place VMCS and VMXON regions in this
area turned out to be too short-sighted. Other file-backed regions would end
up there which destroyed the kernel image layout that assumes the kernel text
region to be the first region at 16#10_0000#.