Suppress warning if CPU_Range only has one element
Turning warnings off completely is not optimal, as we may miss other, serious
issues.  However, if we suppress the warning more specifically (matching the
exact message) this results in a warning that nothing was suppressed if the
range contains more than one element ;-(

TODO: Analyze similar issues with generated types in the code base.