Use preprocessing to selectively enable warnings
In SPARK 2014 the prover relies on the same pragma Warnings mechanism as the
compiler to accept/suppress warnings. As some warnings are only emitted by
later stages of compilation (which are not used for proof) or by the prover it
is not possible to accept all warnings without tripping -gnatw.w "Activate
warnings on Warnings Off pragmas". As an additional issue, -gnatw.w triggers
on the first pass of proof which makes accepting warnings caused by the second
step of proof impossible.

Although any preprocessing may introduce undesired complexity in a program, a
very limited set of preprocessing allows the fine-tuning of warnings for
different modes of compilation:

pragma Warnings is replaced by pragma
 $Compile_Warnings for warnings which occur in release and debug mode only
 $Proof_Warnings   for warnings which occur in prove mode (2nd stage) only
 $Debug_Warnings   for warnings which occur in release mode only
 $Release_Warnings for warnings which occur in release and prove modes only

For compilation active pragmas are changed to "pragma Warnings" whereas
inactive pragmas are changed to the unknown "pragma Ignored_Warnings".
Warnings about "pragma Ignored_Warnings" are suppressed by the configuration file.

As SPARK 2014 does not allow unknown pragmas we have to hijack an existing
pragma which has not side effects. As pragma Debug is ignored and not checked
for content in prove runs, we can transform all instances of inactive warnings
into "pragma Debug".