Libmuxml: Restrict alignment to 4K, 2M or 1G
Alignment is used as a hint for the physical address of a memory region
with the goal to optimize paging structures. It is used by mucfgalloc as a
restriction for placement of the region and mucfcexpand uses the alignment
information for page table region size calculation to enable large page
mappings. Since the expansion step happens before allocation mucfgexpand
has no knowledge of the physical memory addresses.

Restrict the values to the different page sizes to make the calculation
unambiguous. Otherwise a corner case could lead to differences in
calculated PT sizes between the expander and mugenpt (mucfgexpand
assumed large page mappings while mugenpt used 4K pages).

With the now allowed values no such ambiguity is possible.