Add Paging.Pagetables package
The package contains a page table type that stores page table entries.
It is basically a copy of the table-related parts of the Paging.Pagetable
package.

Having a page table that can handle non-specific entries (i.e. not
PDPTE, PDE, etc) is a necessity to simplify the paging code and add
support for 3-level paging structures.

NOTE: The package will be renamed to Paging.Tables once the existing
      code has been reworked and the obsolete code has been removed.