Implement Paging.Layouts.Serialize procedure
The procedure serializes paging structures of a given memory layout.
Pagetables are processed in the order level 1 -> level 4 -> level 3 ->
level 2 using the specified serialization procedures. The provided
serializers are used in ascending order, i.e. Serializers (1) -> level
1, etc.

Reusing the same unit test data as the existing serialization code in
the Paging.Memory package assures that the identical output is
generated.