Implement IOMMU fault interrupt setter
The procedure sets the fault interrupt vector and destination APIC ID of
the specified IOMMU to the given values. An interrupt with the specified
vector is sent to the CPU with the specified APIC ID if the IOMMU
recognizes a primary fault and the fault interrupt mask is not set.