Kernel: Accept warnings caused by systems with no IOMMU
An empty IOMMU device range is generated for systems that have no
IOMMUs and thus the for loops will never be executed. The warning can be
disregarded since this is the desired behavior.

The reason for the warnings pragma is added as a comment since using the
'Reason =>' construct leads to a GNAT box with GNAT GPL 2014.

NOTE: Since the warning is only generated for systems without IOMMUs
      only turning off the specific warning is not possible because it
      will in turn lead to the output of a "no warning suppressed by
      this pragma" warning.