Mucfgexpand: Optimize Calculate_* util functions
Avoid XPath lookups in each iteration of the virtual memory region loop.
Store all physical memory regions prior to entering the loop and use the
new Get_Attribute function to get the size of the corresponding physical
memory region.

Also reduce the complexity of the XPath used to lookup device memory
size. Store all physical devices prior to entering the loop and use the
new the Get_Element function to find the corresponding physical device.

This significantly reduces the number of XPath/DOM traversals and result
in a massive speed up (factor ~8) of the expander tool.

A nice side-effect is also the equally massive reduction in memory
leakage (factor ~7.5) due to less calls to the XIA/XPath library.