Enable complete proofs if supported by gnatprove
The GPL 2014 gnatprove tool is affected by multiple issues ([N425-012],
[N722-005]) that prevent certain proves to be performed. Only enable
all proofs if the gnatprove tool is unaffected by the bugs.

The release date of the tool is used to determine if all proofs can be
performed. The target date is set according to the release date of the
wavefront from ticket [N722-005] that fixes all necessary issues.