Add VTd_Enabled argument to Setup_IRQ_Routing
If this boolean is True, use the IRQ as destination ID instead of the
physical APIC ID. This ID will be the index into the IRT (with an IRT
entry for each active device IRQ).

Whether VTd is enabled should be read from some kind of feature package
but is hard-coded for now.