Libmuxml: Add <barriers> to majorFrameType in format A,B
The new barriers element will be used to specify the barrier
configuration for a given major frame. The config declares a barrier for
each synchronization point of a major frame and how many CPUs must wait
on each barrier.

These elements are optional for now to allow an iterative
implementation.