Mugenspec: Add major frame info types to Skp.Scheduling
The new types currently wrap the major frame barrier config in a record
type. It will be used to consolidate information related to major frames
(barrier config and major frame period in CPU cycles).

The separate barrier config is left as-is to enable an iterative
implementation. It will be removed as soon as the kernel has been
switched to the new major frame info data.