Kernel: Add Scheduler.Set_VMX_Exit_Timer procedure
The procedure sets the VMX-preemption timer by calculating the time
difference from now to the current deadline. If the deadline is already
in the past the timer is set to zero.

The deadline is calculated as follows:

  Deadline := Major_Frame_Start + Current_Minor_Frame.Deadline;

The deadline of the current minor frame is the number of CPU cycles from
the beginning of the major to the end of the given minor frame.