Kernel: Use updated major frame to get next subject ID
When identifying if a new VMCS must be loaded on scheduling frame
change, use the (potentially) updated major frame ID from CPU_Global to
lookup the ID of the next subject that will be run. The stack variable
'Current_Major_ID' is switched to the next major ID on the BSP only.