Globally disable pragma atomic warnings for proofs
The restrictions-prove.adc is only declared as local restriction in
SPARK components. If such a component includes a library, gnatprove will
still issue warnings about unsupported Atomic pragmas for the code
originating from the included project.

Disable these warnings for the complete project hierarchy by moving them
to the global SPARK restrictions.adc file.

For now this is the easiest way to use libdebuglog in both Ada and SPARK
components.