Libpaging: Make Paging_Level a subtype of Positive
This avoids type conversions from the upcoming paging level types
defined in other projects (they are the same thing but independent to
avoid inter-project dependencies).

Update the generated test stub for the Mucfgexpand Calculate_PT_Size
function (it differs because the declaration of the function using the
Paging_Level type has changed).