Kernel: Add implementation for xAPIC mode
If the x2APIC mode is disabled, use the xAPIC mode implementation in the
features/x2apic/disabled directory to communicate with the local APIC.

Access to the APIC registers is modeled using local, volatile variables
instead of a global record type. This saves us from introducing a global
SPARK state and is consistent with how x2APIC mode is modeled (i.e. the
APIC state is part of the X86_64.State).