Add system policy for Genode base-hw x86_64_muen
The genode-base_hw.xml policy specifies a demo system which can be used
to run the Muen port of the Genode base-hw [1] microkernel as subject
(hw_x86_64_muen).

The policy assigns the required resources to the base_hw subject in
order to run the Genode 'run/demo' scenario. This scenario uses VESA
graphics and the nitpicker minimal-complexity GUI server to provide a
graphical environment.

[1] - http://genode.org/documentation/platforms/hw