Kernel: Rename IRQ argument of IO_Apic.Route_IRQ
Rename the IRQ argument to RTE_Index and constrain its type to
Skp.Interrupts.RTE_Index_Type. Use the RTE index stored in the IRQ
routing table instead of the IRQ to configure the I/O APIC. This is
required to make the legacy PIT IRQ0 work which is wired to INTIN2, not
INTIN0.