VT: Implement PS/2 mouse packet processing
The procedure processes the current content of the packet buffer to
generate the corresponding motion input event. The packet header is
inspected to determine if the relative motion values must be
sign-extended by calculating the two's-complement value. The X/Y-axis
data is discarded if a corresponding overflow has been signaled by the
device.

The packet buffer is reset after processing its content.