Kernel: Implement FPU.Has_Valid_State (XSAVE disabled)
If the XSAVE feature is disabled in the policy the FX[SAVE|RSTOR]
instructions are used instead.

The implementation of the FPU.Has_Valid_State function returns True if
SSE/SSE2/SSE3/SSSE3 extensions and support for the FXSAVE and FXRSTOR
instructions are present, see Intel SDM Vol. 3A, chapter 9.6.