Kernel: Add SK.Tau0_Interface package
The package provides the Get_Major_Frame procedure, which returns the
current major frame ID as specified by Tau0. The precondition checks
that the value is only read by the BSP. The BSP is the only core which
takes care of switching major frame IDs if required.

Future extensions of the kernel<->Tau0 interface should be added to this
package.