Policy: Move Linux RAM region mappings
Let Linux RAM memory regions start at 16#0100_0000# which corresponds to
the default value of the CONFIG_PHYSICAL_ALIGN. If the Linux kernel is
configured as relocatable, it will be copied to this address during the
boot process.

With this change, the initramfs mappings are beyond the RAM mappings,
which simplifies enlarging the initramfs regions without affecting other
mappings.