scripts: Add mulog-subject script
This Python script parses log output generated by the dbgserver and only
prints the messages of the subject with the ID specified as command line
parameter. The processing is done on the fly to enable the use of the
tool as a filter, e.g.

  amtterm 192.168.254.1 | mulog-subject.py 0x3

It is intended to be used in conjunction with the Genode autopilot tool,
which automatically builds and executes Genode scenarios and determines
the successful execution by parsing the log output for expected strings.