Libmucfgcheck: Adjust IOMMU_Presence validator
Instead of checking that at least one IOMMU is present if the feature is
active, enforce the presence of two IOMMUs. All systems we have seen so
far were equipped with exactly two IOMMUs, one for the GPU and one
responsible for all other devices.

Therefore this assumption does not impose practically relevant
restrictions and greatly simplifies the generation of the Skp.IOMMU
package.