Libmuxml: Implement Add_Child utility function
The function adds a child with given name to the parent node if it is
missing such an element. The new node is inserted before the first
existing reference node given by name. The element is not added, if no
reference node is found.
The new node is appended to the parent's child list if no reference
names are specified.