Mucfgexpand: Rework subject device BDF assignment
Instead of simply assigning the PCI BDF of the physical device, allocate
the device numbers per subject and set bus and function numbers to 0.
This makes sure that PCI addresses assigned by the BDF expander are all
located on bus 0.

This change now allows to properly handle PCI BDF assignment for logical
devices that are expanded as part of a device class reference. In that
case no explicit PCI address can be specified since a device class may
contain more than a single device.