Enable cspecs generation for Ada components
If an Ada component declares GENERATE_CSPECS in its Makefile, the
mucgenspec tool is called before the actual component build. This allows
the component to add the 'generated' directory to the Source_Dirs
variable and use the component specification constants in the source
code.

Note that changes to the component specification in the policy are
currently not detected automatically. Manually rebuild the component in
this case.