Add optional hashRef element to memory type
A hash reference element will be used to instruct the Mucfgmemhashes
tool to copy the hash element of the referenced memory region after
digest generation.

From an abstract point of view, the element is a way to link multiple
memory regions by declaring that the hash of the content is the same.