Remove atomic constraint from subject interrupts
The pending interrupts data structure is now exclusively accessed by the
kernel executing the corresponding subject. The same holds true for
handling of device IRQs since they are always routed to the CPU of the
subject owning the device.

As there is no concurrent access to the data structure, access does not
need to be atomic and the lock instruction prefix is not required any
longer.