Do not declare Sinfo instance as volatile
Ref T252. The only reason to declare it volatile is because of the
TSC_Schedule_Start and TSC_Schedule_End fields updated by the kernel on
minor frame change. All other fields remain constant.

Handling a volatile object spanning multiple pages is not feasible as it
must be copied onto the stack before passing it to a procedure/function.
As we do not want to drastically increase the component stack size,
treat the sinfo object as constant.

This will be cleaned up in an upcoming iteration by factoring out the
scheduling fields.