Add Loader.Globals package
The package is used to store/retrieve the current target's address space
offset.  It is calculated by subtracting the address of target sinfo
region in the loader address space minus the address of the sinfo region
in the target's address space.

The operations will be used to pass on the current offset between the
memory iteration procedures.