Provide Loader.Process_Target package
The Process procedure of the package takes the address of a target sinfo
memory region as argument. It first checks that the given address points
to a valid sinfo object. It then calculates the offset of the target
address space and calls the internal Process_Memregion procedure for
each target memory region defined in the sinfo object.

If a region is writable and contains a file or fill content, the memory
region is reset to its initial state using the Addrspace package.