Add dummy serial controller to NUC 6i7KYK platform
For now a device with an I/O port resource is required for use as a
kernel diagnostics device. Add a fake serial in order to enable the use
of the Intel Skull Canyon NUC platform. This can be dropped once the
kernel diagnostics device mechanism has been changed to allow building
systems without diagnostics devices.