Introduce interrupt tables manager
The SK.Interrupt_Tables package provides the Manager_Type which is
responsible to setup all required tables (i.e. GDT, IDT, TSS) to handle
interrupts/exceptions using a separate interrupt stack.

Making the manager a type instead of implementing the functionality in a
package gives us the flexibility to place the type instance anywhere in
memory.