Set init_size field in ZP
The init_size field of the Linux boot protocol specifies the linear
memory required during initialization. Set this field to the size of the
available memory from kernel load address to the end of the region. This
is necessary because Linux upstream commit 974f221... (>v4.6-rc5) moves
the compressed kernel to the end of the decompression buffer.

The new code uses the init_size parameter to calculate the memory
location of where to put the compressed kernel. The effect of setting
init_size to match the region of the kernel load address encompassing
region is that the compressed kernel will be copied to the end of the
mapped region. This assures that there is no overlap between compressed
and decompressed kernel.