Kernel: Drop CR4.SMXE check in System_State.Is_Valid
This bit does not indicate whether the host is in SMX mode of operation.
As it turns out, there is no way for the host OS or hypervisor to detect
if SMX is enabled.

We encountered this bit set on a Lenovo T430s booted via PXE and morbo.
Why it was set remains unclear. The bit is only the first step in
entering SMX, see Intel SDM Vol. 2D, section 6.2.1.

Since having this bit set has no side-effect on 'VMX outside of SMX'
mode of operation, drop check for unset CR4.SMXE bit in
System_State.Is_Valid.

If Muen is executed inside SMX and the IA32_FEATURE_CONTROL MSR is not
setup correctly, a #GP would occur. Please note that running Muen under
SMX is not supported and might not work.