Kernel: Panic on VM exits due to NMI or #MC
If a Non-maskable interrupt (NMI) or machine-check exception (#MC)
occurs and the host is in non-root mode, a VM exit occurs with basic
exit reason set to exception or NMI. Panic in this case instead of
looking for a configured trap handler.

These events indicate a serious hardware problem and we must initiate a
controlled restart via the crash audit mechanism in the future.