Implement Crash_Audit.Allocate procedure
The procedure allocates a new crash audit entry. If this operation fails
because no crash audit entries are available, the calling CPU will be
halted.

Internally, a slot index is incremented atomically and assigned to the
entry type if successful. A client is supposed to use the entry type to
access crash audit slots.