Kernel: Declare crash dump instance
Specify Instance variable of type Dump_Type with address and size from
Skp.Kernel package. Since multiple CPUs might access the structure
simultaneously, it is declared as volatile.

The Size aspect guarantees that the user-specified size of the physical
crash dump region is big enough, as otherwise the kernel build would
fail with a meaningful error.