Provide SK.Strings package in common/strings dir
The package provides Img() functions to convert Byte and Word* types to
strings in hexadecimal format. It also provides the Img_Dec function to
convert Word64 to a string in decimal format.

The package is basically a copy of the SK.Utils package with the
addition of the Img_Dec function, which is a copy of Put_UInt64 from
Debuglog.Client.

The package is in SPARK 2014.