Call Dump.Print_VMX_Error from VMX.Error procedure
Make VT-x context an argument of Dump.Print_VMX_Error and call it from
VMX.Error instead of VMX.VMX_Error. Note that the currently active
subject ID is no longer printed as this introduces an unwanted
dependency to the Scheduler package. Furthermore, the information can be
deduced from the active VMCS address.