Implement initial PCI config space emulation
The Devices.Pciconf package provides the Mediate procedure which can be
used to restrict access to the PCI config space of a device.

Currently, read access is allowed while write access is only permitted
for fields which have an explicit write configuration specified:
- Command
- Cache Line Size
- Latency Timer
- BIST
- BAR*

Write requests which do not have a write config are ignored.

PCI capabilities are virtualized so that only MSI/MSI-X are exposed.
Unfiltered Write access is currently granted for both caps. 32-bit MSI,
64-bit MSI and MSI-X caps are working, MSI per-vector masking is
untested due to lack of test hardware.