Kernel: Disable axiom guards in gnatprove
SPARK v18 introduces guards during VC generation that get in the way
of interactive proving (and sometimes slows down automatic proving as
well).

Use gnatprove's release date to determine if axiom guards are
supported. The target date is set to the commit date where axiom
guards where introduced in SPARK, according to the commit history in
the public github repo.