Add Devices.Pciconf.Addrspace package
The package provides Read and Write operations to the PCI config space
of a device given by SID.

In contrast to the [Read|Write]_Config operations in the parent package,
access to memory is restricted. The package overlays the PCI config
space memory with a byte array which is indexed by SID. Other memory is
not accessible.