Implement Mudm.Client package
The package provides the Pciconf_Emulate_[Read|Write] procedures which
handle request/response processing and hand-overs to DM.

The packages should be part of Libmudm, but this is currently not
feasible because it would require SK.Hypercall to be part of a library
project.

Also, component specs should allow the definition of non-channel related
events to have constants in the code.

Since SM is currently the only client, it is ok to have this package
here for now.