Integrate DM into example systems
The global configuration variable 'pciconf_emulation_enabled' controls
whether the PCI configuration space emulation feature is enabled or not.
If it is disabled, the DM subjects with all related structures
(channels, events etc) are not included in the policy.

The 'pciconf_emulation_xhci_enabled' flag controls whether storage SM
uses storage DM to emulate the configuration space of the xHCI
controller (this is required for integration_tests_kt.xml).