Kernel: Adjust event nr. handling in Handle_Hypercall
As it turns out, the conditional execution of code depending on the
user-provided Event_Nr variable might be problematic. This is due to the
recently discovered speculative execution bugs, i.e. bounds check bypass
(CVE-2017-5753), also called variant 1.

Whether this call site is actually exploitable is not definitely clear
at this point but as a precaution we mitigate the potential issue by
avoiding the bounds check via conditional if. Instead, the user-supplied
event number is first transposed to the valid event range via bitwise
and then only the checked value is used as index into arrays.