Dbgserver: Add PC speaker sink
The debug server is extended by a sink which utilizes the PC speaker
for output, using frequency modulation.  Even- and odd-numbered bits
are modulated using distinct frequencies to enable synchronization;
moreover, after each byte a byte seperator frequency is output.

It is recommended to use an audio cable to eliminate disturbing noise.
The output is to be decoded using the upcoming tool mupcspkrdbg.