Rework Sinfo data structures
Use an array of variant records to export memory, event, vector and
device resources to subjects.

The variants are explicitly padded to the size of the largest variant
(Memregion_Type) in order to guarantee proper zeroization of unused
space in smaller variants.

The channel abstraction has been discontinued in favor of additional
vector and event resources. This will allow subjects to query
information about non-channel related vectors and events.

In order to avoid No_Implicit_Loops warnings, the generation of
initialization code for the variant array type is suppressed via
Suppress_Initialization aspect. An explicit Null_Subject_Info constant
is used for initialization of the sinfo object to be serialized by the
Mugensinfo tool. Note that removing the No_Implicit_Loops pragma had the
undesired side effect of stack usage above the allowed limits in
component code.

Increase sinfo magic number in order to reflect the incompatible data
structure changes.