Improve Travis CI configuration
- Enable ccache in docker containers
- Build docker images
- Use pulled image (if any) as image cache
- Push resulting images if branch is master or devel

This fully automates the creation of Docker Hub images for devel and
master branches.

The Muen repository submodules are shallow clones in order to (greatly)
reduce time and space to clone and save them. These repositories can
easily be converted to complete ones by the git fetch --unshallow
command. Note that at least Debian Stretch (9) or Ubuntu Bionic Beaver
(18.04 LTS) is required for this feature to work, as otherwise
the git command is too old and does not support --shallow-submodules.