Adjust stackcheck targets
In order to run these targets in parallel, they must depend on the
associated binary. Since this dependency is phony, a stackcheck target
also becomes a phony target. Therefore, drop stackcheck stamp files as
they do no longer work.

As a result of this change, stackcheck targets are now always executed
when present in a Makefile. As they are quite fast, this should be ok.