Split tau0_static from components build target
Add a separate tau0_static target to the toplevel Makefile and let pack
depend on it. Then, remove tau0_static from components build target.
This enables the use the COMPONENTS environment variable in combination
with the components target again to only build the specified components
without pulling in tau0_static.